Fix IPv6 set and DHCP IP change on config change in NM. #PSBM-45066
IPv6 didn't set in Fedora 23 because NM uses configuration created by ifcfg-rh plugin.
To prevent it need to create NM configuration to disable default config and plugin for interface before configure this interface.
Another problem is change the IP from DHCP after interface configuration change change.
To fix it need to set mac to dhcp-client-id in the configuration.
And to be sure that correct configuration used need to up it for interface.