Clear old NM config if network interface were recreated. #PSBM-48328
Before the fix, if recreate network interface (delete and create an interface with same name) tools didn't delete the old config. It leads to configuration to not exists interface, which block configuration of new one.
Fix add a check on interface change. If it changes, script erases old config and create new for the new interface.