Revert rtnl_dump_filter wrapper
This reverts commit 06de11ff8fc1642fc9f4fd7a996468e5b7589500
"Add precompile if to call rtnl_dump_filter with correct arg number for
libnetlink versions. #PSBM-44221"
Reason: The condition macro does not correctly indicate the variant of code.
In CentOS 6 compilation fails because the function contains 3 arguments, but
the macro is not present.
All actual systems switched to the 3-arg variant long ago, so the 5-arg can be
simply removed.